+++ CHANGELOG.md
@@ -26,6 +26,7 @@
     * Update link to coffee2code.com to be HTTPS
     * Update `readme_url()` to refer to plugin's readme.txt on plugins.svn.wordpress.org
     * Remove defunct line of code
+* Change: Move translation of all parent class strings into main plugin file
 * Change: Note compatibility through WP 5.7+
 * Change: Update copyright date (2021)